So I had this amazing idea
I changed the localhost/e-commerce-store directory in apache to set it
as default and now when we go to localhost it gives us the store. Which
is amazing because before we have to go to localhost/e-commerce-store to
get the site. This emulated a sort of a sub-domain type of thing. This
was fucking all the paths up. I can now finally fix the paths. Lets gets
started.